requirements.txt
streamlit
openai-whisper
transformers
torchaudio
sentencepiece
